helpers contains setting up tools
helpers mod contains functions and structs needed to setup common inputs/outputs

Closes #26
