+++ bmap-rs/Cargo.toml
@@ -10,5 +10,3 @@ edition = "2018"
 bmap = { path = "../bmap" }
 anyhow = "1"
 structopt = "0.3"
-nix = "0.25"
-flate2 = "1"
