+++ bmap-rs/src/main.rs
@@ -1,12 +1,9 @@
 use anyhow::{anyhow, bail, Context, Result};
-use bmap::{Bmap, Discarder, SeekForward};
-use flate2::read::GzDecoder;
-use nix::unistd::ftruncate;
-use std::ffi::OsStr;
+use bmap::helpers::{find_bmap, setup_input, setup_output};
+use bmap::Bmap;
 use std::fs::File;
 use std::io::Read;
-use std::os::unix::io::AsRawFd;
-use std::path::{Path, PathBuf};
+use std::path::PathBuf;
 use structopt::StructOpt;
 
 #[derive(StructOpt, Debug)]
@@ -26,66 +23,6 @@ struct Opts {
     command: Command,
 }
 
-fn append(path: PathBuf) -> PathBuf {
-    let mut p = path.into_os_string();
-    p.push(".bmap");
-    p.into()
-}
-
-fn find_bmap(img: &Path) -> Option<PathBuf> {
-    let mut bmap = img.to_path_buf();
-    loop {
-        bmap = append(bmap);
-        if bmap.exists() {
-            return Some(bmap);
-        }
-
-        // Drop .bmap
-        bmap.set_extension("");
-        bmap.extension()?;
-        // Drop existing orignal extension part
-        bmap.set_extension("");
-    }
-}
-
-trait ReadSeekForward: SeekForward + Read {}
-impl<T: Read + SeekForward> ReadSeekForward for T {}
-
-struct Decoder {
-    inner: Box<dyn ReadSeekForward>,
-}
-
-impl Decoder {
-    fn new<T: ReadSeekForward + 'static>(inner: T) -> Self {
-        Self {
-            inner: Box::new(inner),
-        }
-    }
-}
-
-impl Read for Decoder {
-    fn read(&mut self, data: &mut [u8]) -> std::io::Result<usize> {
-        self.inner.read(data)
-    }
-}
-
-impl SeekForward for Decoder {
-    fn seek_forward(&mut self, forward: u64) -> std::io::Result<()> {
-        self.inner.seek_forward(forward)
-    }
-}
-
-fn setup_input(path: &Path) -> Result<Decoder> {
-    let f = File::open(path)?;
-    match path.extension().and_then(OsStr::to_str) {
-        Some("gz") => {
-            let gz = GzDecoder::new(f);
-            Ok(Decoder::new(Discarder::new(gz)))
-        }
-        _ => Ok(Decoder::new(f)),
-    }
-}
-
 fn copy(c: Copy) -> Result<()> {
     if !c.image.exists() {
         bail!("Image file doesn't exist")
@@ -99,13 +36,7 @@ fn copy(c: Copy) -> Result<()> {
     b.read_to_string(&mut xml)?;
 
     let bmap = Bmap::from_xml(&xml)?;
-    let mut output = std::fs::OpenOptions::new()
-        .write(true)
-        .create(true)
-        .open(c.dest)?;
-
-    ftruncate(output.as_raw_fd(), bmap.image_size() as i64).context("Failed to truncate file")?;
-
+    let mut output = setup_output(&c.dest, &bmap)?;
     let mut input = setup_input(&c.image)?;
     bmap::copy(&mut input, &mut output, &bmap)?;
     println!("Done: Syncing...");
