+++ bmap/src/helpers.rs
@@ -1,0 +1,78 @@
+use crate::{Bmap, Discarder, SeekForward};
+use anyhow::{Context, Result};
+use flate2::read::GzDecoder;
+use nix::unistd::ftruncate;
+use std::ffi::OsStr;
+use std::fs::File;
+use std::io::Read;
+use std::os::unix::io::AsRawFd;
+use std::path::{Path, PathBuf};
+
+pub fn append(path: PathBuf) -> PathBuf {
+    let mut p = path.into_os_string();
+    p.push(".bmap");
+    p.into()
+}
+
+pub fn find_bmap(img: &Path) -> Option<PathBuf> {
+    let mut bmap = img.to_path_buf();
+    loop {
+        bmap = append(bmap);
+        if bmap.exists() {
+            return Some(bmap);
+        }
+
+        // Drop .bmap
+        bmap.set_extension("");
+        bmap.extension()?;
+        // Drop existing orignal extension part
+        bmap.set_extension("");
+    }
+}
+
+trait ReadSeekForward: SeekForward + Read {}
+impl<T: Read + SeekForward> ReadSeekForward for T {}
+
+pub struct Decoder {
+    inner: Box<dyn ReadSeekForward>,
+}
+
+impl Decoder {
+    fn new<T: ReadSeekForward + 'static>(inner: T) -> Self {
+        Self {
+            inner: Box::new(inner),
+        }
+    }
+}
+
+impl Read for Decoder {
+    fn read(&mut self, data: &mut [u8]) -> std::io::Result<usize> {
+        self.inner.read(data)
+    }
+}
+
+impl SeekForward for Decoder {
+    fn seek_forward(&mut self, forward: u64) -> std::io::Result<()> {
+        self.inner.seek_forward(forward)
+    }
+}
+
+pub fn setup_input(path: &Path) -> Result<Decoder> {
+    let f = File::open(path)?;
+    match path.extension().and_then(OsStr::to_str) {
+        Some("gz") => {
+            let gz = GzDecoder::new(f);
+            Ok(Decoder::new(Discarder::new(gz)))
+        }
+        _ => Ok(Decoder::new(f)),
+    }
+}
+
+pub fn setup_output(path: &Path, bmap: &Bmap) -> Result<File> {
+    let output = std::fs::OpenOptions::new()
+        .write(true)
+        .create(true)
+        .open(path)?;
+    ftruncate(output.as_raw_fd(), bmap.image_size() as i64).context("Failed to truncate file")?;
+    Ok(output)
+}
