+++ bmap/src/lib.rs
@@ -4,6 +4,7 @@ mod discarder;
 pub use crate::discarder::*;
 use sha2::{Digest, Sha256};
 use thiserror::Error;
+pub mod helpers;
 
 use std::io::Result as IOResult;
 use std::io::{Read, Seek, SeekFrom, Write};
